+++ audiomate/corpus/io/__init__.py
@@ -29,6 +29,7 @@ from .rouen import RouenDownloader, RouenReader  # noqa: F401
 from .audio_mnist import AudioMNISTDownloader, AudioMNISTReader  # noqa: F401
 from .fluent_speech import FluentSpeechReader  # noqa: F401
 from .wav2letter import Wav2LetterWriter  # noqa: F401
+from .nvidia_jasper import NvidiaJasperWriter  # noqa: F401
 
 __downloaders = {}
 for cls in CorpusDownloader.__subclasses__():